refactor: extract resume page inline components into separate files
Split ResumeProfile, ResumeContent, and ContentContainer out of
app/resume/page.tsx into app/resume/components/ for better
maintainability and readability.
